refactored the report summary logic to determine census counts using data already in the report service
