+++ Web/Admin.UI/src/app/components/tenant/facility-view/facility-view.component.scss
@@ -39,3 +39,12 @@ $primary: #398eb4;
     margin-right: 5px;
     display: inline-block;
 }
+
+.adhoc-chip {
+    background-color: #43007D; 
+    color: #FFFFFF;
+    padding: 5px 10px;
+    border-radius: 16px;
+    margin-right: 5px;
+    display: inline-block;
+}
